--- a/src/Control/Egison/Match.hs
+++ b/src/Control/Egison/Match.hs
@@ -15,6 +15,7 @@ import           Data.Type.Equality
 -- | @matchAll@ takes a target, a matcher, and a list of match clauses.
 -- @matchAll@ collects all the pattern-matching results and returns a list of the results evaluating the body expression for each pattern-matching result.
 -- @matchAll@ traverses a search tree for pattern matching in breadth-first order.
+{-# INLINE matchAll #-}
 matchAll :: (Matcher m a) => a -> m -> [MatchClause a m b] -> [b]
 matchAll tgt m = foldr go []
   where
@@ -25,10 +26,12 @@ matchAll tgt m = foldr go []
 -- | @match@ takes a target, a matcher, and a list of match clauses.
 -- @match@ calculates only the first pattern-matching result and returns the results evaluating the body expression for the first pattern-matching result.
 -- @match@ traverses a search tree for pattern matching in breadth-first order.
+{-# INLINE match #-}
 match :: (Matcher m a) => a -> m -> [MatchClause a m b] -> b
 match tgt m cs = head $ matchAll tgt m cs
 
 -- | @matchAllDFS@ is much similar to @matchAll@ but traverses a search tree for pattern matching in depth-first order.
+{-# INLINE matchAllDFS #-}
 matchAllDFS :: (Matcher m a) => a -> m -> [MatchClause a m b] -> [b]
 matchAllDFS tgt m = foldr go []
   where
@@ -37,6 +40,7 @@ matchAllDFS tgt m = foldr go []
       map f results ++ acc
 
 -- | @matchDFS@ is much similar to @match@ but traverses a search tree for pattern matching in depth-first order.
+{-# INLINE matchDFS #-}
 matchDFS :: (Matcher m a) => a -> m -> [MatchClause a m b] -> b
 matchDFS tgt m cs = head $ matchAllDFS tgt m cs
 
@@ -68,6 +72,7 @@ processMStates :: [MState vs] -> [[MState vs]]
 processMStates []          = []
 processMStates (mstate:ms) = [processMState mstate, ms]
 
+{-# INLINE processMState #-}
 processMState :: MState vs -> [MState vs]
 processMState (MState rs (MCons (MAtom pat m tgt) atoms)) =
   case pat of
@@ -88,12 +93,15 @@ processMState (MState rs (MCons (MAtom pat m tgt) atoms)) =
     PredicatePat f -> [MState rs atoms | f rs tgt]
 processMState (MState rs MNil) = undefined -- or [MState rs MNil] -- TODO: shold not reach here but reaches here.
 
+{-# INLINE patVarProof #-}
 patVarProof :: HList xs -> HList '[a] -> MList (xs :++: '[a]) ys -> ((xs :++: '[a]) :++: ys) :~: (xs :++: ('[a] :++: ys))
 patVarProof HNil _ _ = Refl
 patVarProof (HCons _ xs) ys zs = unsafeCoerce Refl -- Todo: Write proof.
 
+{-# INLINE andPatProof #-}
 andPatProof :: MAtom ctx vs -> MAtom (ctx :++: vs) vs' -> MList (ctx :++: vs :++: vs') ys -> (ctx :++: ((vs :++: vs') :++: ys)) :~: (ctx :++: (vs :++: (vs' :++: ys)))
 andPatProof _ _ _ = unsafeCoerce Refl -- Todo: Write proof.
 
+{-# INLINE assocProof #-}
 assocProof :: MAtom ctx vs -> MAtom (ctx :++: vs) vs' -> (ctx :++: (vs :++: vs')) :~: ((ctx :++: vs) :++: vs')
 assocProof _ _ = unsafeCoerce Refl -- Todo: Write proof.