- Added `INLINE` pragmas to some functions
- Enabled `Strict`, `StrictData` extensions
- Reduced the number of iterations  in `processMStatesAll`
- ... and some fixes (7ff0e45a42a1f0b014d31b1cf0321dbc77733924, e8c37a6ff1b65ae2246875fff79662c58a1de465)

```
$ bench ./cdcl_master
benchmarking ./cdcl_master
time                 2.993 s    (2.935 s .. 3.091 s)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 2.960 s    (2.947 s .. 2.979 s)
std dev              18.30 ms   (3.672 ms .. 23.99 ms)
variance introduced by outliers: 19% (moderately inflated)
```

```
$ bench ./cdcl_fast
benchmarking ./cdcl_fast
time                 1.878 s    (1.770 s .. 1.988 s)
                     1.000 R²   (0.998 R² .. 1.000 R²)
mean                 1.903 s    (1.888 s .. 1.930 s)
std dev              26.15 ms   (2.380 ms .. 33.05 ms)
variance introduced by outliers: 19% (moderately inflated)
```